Split frontend Playwright workflows
