+++ frontend/tests/template-login-status.spec.ts
@@ -1,271 +1,16 @@
-import { readFileSync } from "node:fs"
-import { expect, type Locator, type Page, test } from "@playwright/test"
+import { expect, test } from "@playwright/test"
 import { authHeaders, login, testUserEmail } from "./auth-helpers"
 import { evidenceScreenshotPath } from "./evidence-paths"
-
-const validCveList = Buffer.from("CVE-2021-44228\nCVE-2024-3094\n")
-const cyclonedxVex = readFileSync("../data/input_fixtures/cyclonedx_vex.json")
-const validOccurrenceCsv = Buffer.from(
-  [
-    "cve_id,asset_ref,component,version,purl,severity,owner,business_service,exposure",
-    "CVE-2024-3094,build-host-1,xz <img src=x onerror=window.__vpwXss=1>,5.6.0,pkg:apk/alpine/xz@5.6.0-r0,CRITICAL,team-platform <img src=x onerror=window.__vpwXss=1>,payments <script>window.__vpwXss=1</script>,public",
-    "CVE-2024-4577,web-tier,php-cgi,8.3.7,pkg:deb/debian/php-cgi@8.3.7,HIGH,team-web,checkout,internal",
-  ].join("\n"),
-)
-const validAssetContextCsv = Buffer.from(
-  [
-    "target_kind,target_ref,asset_id,owner,business_service,criticality,exposure,environment",
-    "generic,ui-sidecar,ui-sidecar-asset,team-sidecar,inventory-sidecar,high,internal,staging",
-  ].join("\n"),
-)
-const importWizardOpenVex = Buffer.from(
-  JSON.stringify({
-    statements: [
-      {
-        justification: "vulnerable_code_not_present",
-        products: [
-          {
-            identifiers: {
-              purl: "pkg:apk/alpine/xz@5.6.0-r0?arch=x86_64",
-            },
-          },
-        ],
-        status: "not_affected",
-        vulnerability: { name: "CVE-2024-3094" },
-      },
-    ],
-  }),
-)
-const invalidOccurrenceCsv = Buffer.from(
-  [
-    "cve_id,asset_ref,component_name,component_version,purl,scanner,fix_version,severity,owner,business_service,exposure",
-    "not-a-cve,build-host-1,xz,5.6.0,pkg:apk/alpine/xz@5.6.0-r0,trivy,5.6.1-r2,CRITICAL,team-platform,payments,public",
-  ].join("\n"),
-)
-const cyclonedxVexOccurrenceCsv = Buffer.from(
-  [
-    "cve_id,asset_ref,component,version,purl,severity",
-    "CVE-2023-34362,moveit-service,moveit-transfer,2023.0.0,pkg:pypi/moveit-transfer@2023.0.0,HIGH",
-  ].join("\n"),
-)
-
-async function selectRadixOption(
-  page: Page,
-  trigger: Locator,
-  optionName: string | RegExp,
-) {
-  await expect(trigger).toBeVisible()
-  await trigger.click()
-  if (typeof optionName === "string") {
-    const option = page.getByRole("option", { exact: true, name: optionName })
-    await expect(option).toBeAttached()
-    await option.evaluate((element) => {
-      const target = element as HTMLElement
-      let parent = target.parentElement
-      while (parent) {
-        if (parent.scrollHeight > parent.clientHeight) {
-          parent.scrollTop = target.offsetTop - parent.clientHeight / 2
-          break
-        }
-        parent = parent.parentElement
-      }
-      target.scrollIntoView({ block: "nearest" })
-    })
-    try {
-      await option.click({ timeout: 3000 })
-    } catch {
-      await option.dispatchEvent("pointerdown")
-      await option.dispatchEvent("pointerup")
-      await option.dispatchEvent("click")
-    }
-    return
-  }
-  const option = page.getByRole("option", { name: optionName })
-  try {
-    await option.scrollIntoViewIfNeeded({ timeout: 1000 })
-    await option.click({ timeout: 3000 })
-  } catch {
-    await option.click({ force: true })
-  }
-}
-
-async function selectRadixOptionByLabel(
-  page: Page,
-  scope: Page | Locator,
-  label: string,
-  optionName: string | RegExp,
-) {
-  await selectRadixOption(
-    page,
-    scope.getByRole("combobox", { exact: true, name: label }),
-    optionName,
-  )
-}
-
-async function selectDashboardProject(page: Page, projectName: string) {
-  const projectTrigger = page.getByRole("combobox").first()
-  await selectRadixOption(page, projectTrigger, projectName)
-  await expect(projectTrigger).toContainText(projectName)
-}
-
-test("template finding detail renders TTP Context tab", async ({ page }) => {
-  test.setTimeout(60_000)
-  const testRunSuffix = Date.now().toString(36)
-  const projectName = `VPW TTP Context ${testRunSuffix}`
-
-  const accessToken = await login(page)
-  const headers = authHeaders(accessToken)
-  const projectResponse = await page.request.post(
-    "http://127.0.0.1:8000/api/v1/projects/",
-    {
-      data: {
-        description: "Playwright TTP Context project",
-        name: projectName,
-      },
-      headers,
-    },
-  )
-  expect(projectResponse.ok()).toBeTruthy()
-  const project = (await projectResponse.json()) as { id: string }
-
-  const importResponse = await page.request.post(
-    `http://127.0.0.1:8000/api/v1/projects/${project.id}/imports`,
-    {
-      headers,
-      multipart: {
-        file: {
-          buffer: validCveList,
-          mimeType: "text/plain",
-          name: "cves.txt",
-        },
-        input_type: "cve-list",
-      },
-    },
-  )
-  expect(importResponse.ok()).toBeTruthy()
-
-  const findingsResponse = await page.request.get(
-    `http://127.0.0.1:8000/api/v1/projects/${project.id}/findings/?sort=cve`,
-    { headers },
-  )
-  expect(findingsResponse.ok()).toBeTruthy()
-  const findingsPayload = (await findingsResponse.json()) as {
-    data: Array<{ cve_id: string; id: string }>
-  }
-  const finding = findingsPayload.data.find(
-    (item) => item.cve_id === "CVE-2024-3094",
-  )
-  expect(finding?.id).toBeTruthy()
-  if (!finding?.id) {
-    throw new Error("Expected CVE-2024-3094 finding to exist.")
-  }
-
-  await page.goto("/")
-  await selectDashboardProject(page, projectName)
-  await page.screenshot({
-    fullPage: true,
-    path: evidenceScreenshotPath("vpw-059-attack-summary-dashboard.png"),
-  })
-
-  await page.goto(`/findings/${finding.id}`)
-  await expect(
-    page.getByRole("heading", { name: "CVE-2024-3094" }),
-  ).toBeVisible()
-  await page.getByRole("tab", { name: "TTP Context" }).click()
-  const ttpPanel = page.getByRole("tabpanel", { name: "TTP Context" })
-  await expect(ttpPanel).toContainText("Defensive context")
-  await expect(ttpPanel).toContainText(
-    "No approved ATT&CK mapping is stored for this finding.",
-  )
-  await expect(ttpPanel).toContainText(
-    "Workbench does not infer tactics or techniques for unmapped CVEs.",
-  )
-  await page.screenshot({
-    fullPage: true,
-    path: evidenceScreenshotPath("vpw-058-ttp-context-tab.png"),
-  })
-})
-test("template frontend renders CycloneDX VEX occurrence evidence", async ({
-  page,
-}) => {
-  test.setTimeout(60_000)
-  const testRunSuffix = Date.now().toString(36)
-  const projectName = `VPW CycloneDX VEX ${testRunSuffix}`
-
-  const accessToken = await login(page)
-  const headers = authHeaders(accessToken)
-  const projectResponse = await page.request.post(
-    "http://127.0.0.1:8000/api/v1/projects/",
-    {
-      data: {
-        description: "Playwright CycloneDX VEX project",
-        name: projectName,
-      },
-      headers,
-    },
-  )
-  expect(projectResponse.ok()).toBeTruthy()
-  const project = (await projectResponse.json()) as { id: string }
-
-  const importResponse = await page.request.post(
-    `http://127.0.0.1:8000/api/v1/projects/${project.id}/imports`,
-    {
-      headers,
-      multipart: {
-        file: {
-          buffer: cyclonedxVexOccurrenceCsv,
-          mimeType: "text/csv",
-          name: "cyclonedx-vex-occurrence.csv",
-        },
-        input_type: "generic-occurrence-csv",
-        vex_file: {
-          buffer: cyclonedxVex,
-          mimeType: "application/json",
-          name: "cyclonedx-vex.json",
-        },
-      },
-    },
-  )
-  expect(importResponse.ok()).toBeTruthy()
-
-  const findingsResponse = await page.request.get(
-    `http://127.0.0.1:8000/api/v1/projects/${project.id}/findings/?sort=cve`,
-    { headers },
-  )
-  expect(findingsResponse.ok()).toBeTruthy()
-  const findingsPayload = (await findingsResponse.json()) as {
-    data: Array<{
-      cve_id: string
-      id: string
-      status?: string
-      suppressed_by_vex?: boolean
-    }>
-  }
-  const suppressed = findingsPayload.data.find(
-    (finding) => finding.cve_id === "CVE-2023-34362",
-  )
-  expect(suppressed).toBeTruthy()
-  if (!suppressed) {
-    throw new Error("Expected CycloneDX VEX-suppressed finding.")
-  }
-  expect(suppressed.suppressed_by_vex).toBe(true)
-  expect(suppressed.status).toBe("suppressed")
-  await page.goto(`/findings/${suppressed.id}`)
-  const occurrencesTable = page.getByRole("table", {
-    name: "Occurrences table",
-  })
-  await expect(occurrencesTable).toContainText("Not Affected")
-  await expect(occurrencesTable).not.toContainText(
-    "vulnerable_code_not_present",
-  )
-  await expect(occurrencesTable).toContainText(
-    "pkg:pypi/moveit-transfer@2023.0.0",
-  )
-  await page.screenshot({
-    fullPage: true,
-    path: evidenceScreenshotPath("vpw-066-cyclonedx-vex-parser.png"),
-  })
-})
+import {
+  importWizardOpenVex,
+  invalidOccurrenceCsv,
+  selectDashboardProject,
+  selectRadixOption,
+  selectRadixOptionByLabel,
+  validAssetContextCsv,
+  validCveList,
+  validOccurrenceCsv,
+} from "./workbench-e2e-helpers"
 test("template frontend covers core Workbench E2E smoke", async ({ page }) => {
   test.setTimeout(180_000)
   const testRunSuffix = Date.now().toString(36)
@@ -959,43 +704,4 @@ test("template frontend covers core Workbench E2E smoke", async ({ page }) => {
   await page.getByRole("menuitem", { name: "Sign out" }).click()
   await expect(page).toHaveURL(/\/login$/)
   await expect(page.getByRole("heading", { name: "Sign in" })).toBeVisible()
-})
-
-test("template settings clears one-time API token when leaving settings", async ({
-  page,
-}) => {
-  const testRunSuffix = Date.now().toString(36)
-
-  await login(page)
-
-  await page.goto("/settings")
-  await expect(
-    page.getByRole("heading", { name: "Service Token" }),
-  ).toBeVisible()
-  await page.getByLabel("Name").fill(`automation-${testRunSuffix}`)
-  const importScope = page.getByRole("checkbox", { name: /IMPORT/i })
-  await importScope.focus()
-  await page.keyboard.press("Space")
-  await expect(importScope).toBeChecked()
-  const reportScope = page.getByRole("checkbox", { name: /REPORT/i })
-  await reportScope.focus()
-  await page.keyboard.press("Space")
-  await expect(reportScope).toBeChecked()
-  await page.getByRole("button", { name: "Create Token" }).click()
-
-  const createdTokenPanel = page.getByRole("region", {
-    name: "Created API token",
-  })
-  await expect(createdTokenPanel).toBeVisible()
-  await expect(createdTokenPanel.getByLabel("Token")).toHaveValue(/^vpr_/)
-  await expect(createdTokenPanel).toContainText("READ, IMPORT, REPORT")
-
-  await page.getByRole("link", { name: "Dashboard" }).click()
-  await page.getByRole("link", { name: "Settings" }).click()
-
-  await expect(createdTokenPanel).toHaveCount(0)
-  await expect(page.getByRole("textbox", { name: "Token" })).toHaveCount(0)
-  await expect(
-    page.getByRole("table", { name: "API tokens table" }),
-  ).toContainText(/READ\s*IMPORT\s*REPORT/)
 })
